+++ CHANGELOG.md
@@ -2,9 +2,19 @@
 
 This project follows [Semantic Versioning](http://semver.org/).
 
-## [36.21.0] 2025-06-27
+## [36.21.0] 2025-07-11
 ### Added
 - Support for observationExtract valueCode relationships (component or member).
+
+## [36.20.3] 2025-07-10
+### Fixed
+- Fixed a bug that the matrix layout ('table' in itemControl extension) 
+  does not always display answers loaded from answerValueSet.
+  
+## [36.20.2] 2025-07-09
+### Fixed
+- answerValueSet expansion error under the field should still show
+  when the form is loaded for the second time.
 
 ## [36.20.1] 2025-06-20
 ### Changed